requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0